Consolidate the remainder of the configuration API into a single module
